vm_tools: Convert HandleStatefulUpdate to async
The blocking version of this call caused an error where concierge would
be permanently unresponsive.

The HandleStatefulUpdate heartbeat occurs every 10 seconds and the grpc
timeout is also 10 seconds, so 100% of concierge's time would be spent
handling a blocking call.

We fix this by converting the RPC to an async variant.

Note that it is a separate problem that VMs become unresponsive, this
only fixes the fact that an unresponseive VM will cause concierge to
also be unresponsive.

BUG=b:289456807
TEST=inspected perfetto trace for an unresponsive VM.

Change-Id: I7cfd318a4ec5e82af5cd9732e207f0a6e28b6c48
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4787832
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Nic Hollingum <hollingum@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>