net-base: update the returned type of Socket::Ioctl() method
According to the description of ioctl() function:
> Usually, on success zero is returned.  A few ioctl() requests use
> the return value as an output parameter and return a nonnegative
> value on success.  On error, -1 is returned, and errno is set to
> indicate the error.

We should pass the nonnegative value back to caller, and convert the
negative value as failure. This CL changes the returned type of
Socket::Ioctl() to std::optional<int>.

Also, it adds the comment for each Socket's method that the caller
should use PLOG to print errno on failure case.

BUG=b:295263596
TEST=FEATURES=test emerge-brya net-base shill

Change-Id: I2e49ca99d0dca3f9475175919abc9763e2bfcd93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4780263
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>