rmad: Add rpc_utils for helper functions to loop RPC calls.
Add a helper function to loop through an async function call with a
mechanism to continue or break with the current input. Can be used in
scenarios like finding the first external storage that contains a file,
where the mount operation might be async.

BUG=b:294012641
TEST=FEATURES=test emerge-dedede rmad

Change-Id: I4956d97526b01adf154d87e33036802ae5f2c159
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4787475
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Jeff Lin <jeffulin@google.com>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>