flex_hwis: Check before calling value on optional
Calling value() on a std::optional can raise a bad_optional_access
exception. We don't want to crash if we can't get a uuid, so we should
check before retrieving the value.

FlexHwisMojo::SetHwisUuid just wraps assignment to a field of a data
structure not owned by FlexHwisMojo. It provides little benefit and
obscures the possible nullopt being passed. It also doesn't quite go
with the other member functions of FlexHwisMojo, which are about mojo to
proto conversion.

Remove FlexHwisMojo::SetHwisUuid, and handle the possibility of a
nullopt in FlexHwisSender, where we're in a better position to handle
it.

BUG=None
TEST=cros_workon start -b reven chromeos-base/flex_hwis && \
     cros_workon_make -b reven --test chromeos-base/flex_hwis

Change-Id: Ia9b8667a05f721260258887dbd8f56d87f19da20
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4784266
Tested-by: Ted Brandston <tbrandston@google.com>
Auto-Submit: Ted Brandston <tbrandston@google.com>
Reviewed-by: Ting-Hao Lin <tinghaolin@google.com>
Commit-Queue: Ting-Hao Lin <tinghaolin@google.com>