shill: WiFi: Remove the check on the time interval in WiFiLinkStatisticsTest
WiFiLinkStatisticsTest tests match the log strings including the
time interval between the start and end network events, assuming the
value is always 0 in unit tests. This hard-coded value makes the test
flaky, as the actual time interval may be nonzero on slow machines.

Remove the check on the time interval to stabilize the unit test, as the
goal of the unit test is to verify the difference of traffic count at
the start and end network events.

BUG=b:296159426
TEST=FEATURES=test emerge-$BOARD shill, manually change the time
intervals in WiFiLinkStatistics::UpdateNl80211LinkStatistics and
WiFiLinkStatistics::UpdateRtnlLinkStatistics.

Change-Id: I470430bf180f7e35eb047d57799801ef42f66295
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4785692
Commit-Queue: Kaidong Wang <kaidong@google.com>
Tested-by: Kaidong Wang <kaidong@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>