cryptohome: Rename UserSecretStashStorage -> UssStorage
While in general abbreviations are supposed to be avoided, USS is a
pretty pervasive one and User Secret Stash is just a little too verbose
to work well, especially once you need to start adding additional words
onto the type and function names.

Using "Uss" is already pretty pervasive, except for the UserSecretStash
and UserSecretStashStorage classes themselves. The former should be
fixed up during a future refactor of that class so I'm not going to
worry about it but the latter is a quick and easy rename.

BUG=b:270621807
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I413bac3018bafe777068ba03bd58a66a060f01c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4786548
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Betul Soysal <betuls@google.com>