patchpanel: Define an interface for CLAT service
This CL adds the interface of a new service "ClatService".
The service subscribes to OnShillLogicalDeviceChanged and
OnIPConfigsChanged, judging whether CLAT is needed or not.
It starts or stops CLAT based on the judgement.

Implementation is not included in this CL, and will be
made in later CLs.

BUG=b:278970851
TEST=cros_workon_make --board=guybrush chromeos-base/patchpanel --test

Change-Id: Ib7a4e259e8a71a78c2cf1c6d82c97fa55221d293
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4792271
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Naoya Tezuka <naoyatezuka@google.com>
Reviewed-by: Cassie Wang <cassiewang@google.com>
Tested-by: Naoya Tezuka <naoyatezuka@google.com>