cryptohome: Implement RelabelAuthFactor
The implementation is only done for USS-backed factors, trying to
relabel a VK factor is treated as an error. Loading and resaving a vault
keyset with a new label is in theory easy (it's just one file) but
saving a VK that you haven't decrypted is a bit unclear to me.

We can still add in support later if it turns out this is a problem.
Unfortunately unlike with most of the other mutating operations we can't
just auto-migrate the factor since a relabel doesn't take the
credentials needed to do so as an input.

BUG=b:292543364
TEST=FEATURES=test emerge-${BOARD} cryptohome

Change-Id: I3567b8553124a5bf8f263426fa7e5953958d7dc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4795448
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>