hammerd: Add retry to reduce false alarm of pairing challenge failure
Users could disconnect the base keyboard while hammerd is conducting the
pairing challenge, thus the USB transfer fail and hammerd then send DBus
message to warn user the keyboard is malicious. Though in the existing
implementation, hammerd would check USB existence and do nothing if
hammerd found the USB is disconnected, the USB endpoint may be present
because if user re-attach the keyboard quickly, or OS not yet remove the
endpoint, so there are still quite a chance user get the false alarm of
connecting a different, malicious keyboard.

Add a retry mechanism for failed USB transfer to further reduce the
chance of getting the false alarm, and only if all retry fail, hammerd
then triggers the DBus to notify user it is connecting to a different
keyboard.

BUG=b:294180296
TEST=(1) CQ
     (2) Deploy to a DUT and continuously attach and reattach the
         keyboard.
     (3) cros_run_unit_tests --emerge-verbose --board staryu --packages
         'chromeos-base/hammerd'

Change-Id: I6d828e11f2ebf9a0c12679eb31eee2304e1d4776
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4792642
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>