hammerd: Improve fw updater debug message
The USB transfer API could fail and no packets are actually transferred,
thus the first byte of the response array may not be the status code.
Update the logic to only print the status of subcommand only if the
transferred bytes are not zero.

BUG=b:294180296
TEST=CQ

Change-Id: If11a690aaf3770489d43be5960dd01813497db66
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4792643
Reviewed-by: Fei Shao <fshao@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Sung-Chi Li <lschyi@chromium.org>