lorgnette: Wrap libsane calls for testability
Currently, lorgnette calls directly into libsane functions to
communicate with a SANE backend.  This means unit tests have to set up a
real backend.  Even with the existing libsane-test backend, this leaves
a lot of error-handling and less-common paths untested.  To make this
easier to test, wrap all the libsane calls in a thin C++ class that can
be replaced with a fake at testing time.  No behavior changes in this
CL.

The LibsaneWrapperImplFake added here is a minimal implementation to
prove that it can be swapped out in tests.  A more complete
implementation will come in a subsequent CL with the code to be tested.

BUG=b:274860707
TEST=CQ

Change-Id: Ic5f70b0122f93207ea3b7aff42877a18019bb04b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4811024
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Clayton Whitelaw <cawhitelaw@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Nathan Muggli <nmuggli@google.com>