lorgnette: Rename option container-related variables
Lorgnette handles a small set of known options needed for the current
Scan app.  It is going to be handling all options exposed by the SANE
backend, which makes the options_ name confusing.  Rename it to
known_options_ in preparation for the new variables.  No functionality
changes in this CL.

BUG=b:274860707
TEST=CQ

Change-Id: Ic2a9666a8a1256cae13cd789e6df0a6e8a16d5f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4811025
Reviewed-by: Clayton Whitelaw <cawhitelaw@google.com>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>