login: Refactor PolicyService:Store method to use completion only
Previously the method was returning the result through two independent
variables: Completion and the return value from the method.

The method has to be asynchronous in case of DevicePolicyService
because it has to update VPD. Which means, we need completion as
a way to return the value. So we remove the redundancy by removing
the return type of the method.

BUG=b:279727161
TEST=Unit tests pass

Change-Id: I78734f7ba449085cf7864d6503b09d48fff0fb7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4800463
Tested-by: Igor <igorcov@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Igor <igorcov@chromium.org>