shill: cellular: Reset the AutoConnect timer after registration
When a cellular device is registered, reset the auto connect timer to
ensure the AutoConnect is triggered immediately. The UE might end up
with a long AutoConnect timeout when changing the custom APNs in the UI,
or when the carrier disconnects the UE for any reason.

BUG=b:294482332
TEST=-Set a bad DEFAULT APN and good Attach APN using the APN revamp UI
-wait some time and check that the autoconnect timeout increases
-set a bad attach APN to ensure de-registration
-set a good Attach APN and verify that the autoconnect timeout was reset.

Change-Id: I265ba60e2e8379583ff97ebebae1328f03666446
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4807432
Reviewed-by: Madhav <madhavadas@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>