shill: cellular: Trigger an AutoConnect when changing an APN
Currently, if the device is using a bad APN to connect, and the user
sets the correct APN in the UI, the UE does not reconnect immediately
because the AutoConnect timeout has a big value from the previous
connection attemps. Reset the timeout so an AutoConnect is triggered
immendiately when fixing an incorrect APN.

BUG=b:294482332
TEST=Set an incorrect APN in the UI. Let the AutoConnect timeout
increase for some time.
Correct the UI APN. Verify that the connecion is immediate.

Change-Id: Ief29b11d36e7f790c48f85e444bc0fc2267c32cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4807433
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Madhav <madhavadas@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>