cryptohome: Implement ReplaceAuthFactor
Similar to the Relabel implementation, this only supports replacing USS
factors and not VK factors. In theory we can implement this if we really
need to but it's very difficult to do so transactionally because we
would need to do a completely different sequence of operations since we
can't use the "write to USS" as the commit point for this operation.

This also includes support in the CLI for the replace operation. This
was done separately for relabel but it turns out it's somewhat
convenient to have the support enabled as a single unit.

BUG=b:292543364
TEST=FEATURES=test emerge-${BOARD} cryptohome

Change-Id: I9ef081c7076e2f8e4c0f44f25de3ccd0ea882706
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4824590
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>