arc: keymint: Add ConvertEnum for KeyParams
This CL adds a ConvertEnum function for Key Params
specifically. Right now, we were treating Enum Param
as an integer, which worked with Keymaster, but
not with KeyMint.

Without the ConvertEnum function, tag values were not
matching correctly in the Reference KeyMint
implementation leading to invalid tags. This CL
solves the issue.

BUG=b:297132947
TEST=emerge-hatch chromeos-base/arc-keymint
TEST=cros_workon_make --board=hatch chromeos-base/arc-keymint --test

Change-Id: Ie2b20c4dd4c8a1b2b1ad21e14eac847d4c030fed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4807220
Commit-Queue: Vaibhav Raheja <vraheja@chromium.org>
Tested-by: Vaibhav Raheja <vraheja@chromium.org>
Reviewed-by: Yao Li <yaohuali@google.com>