shill: cellular: Don't CHECK for empty APN list
Because the APN list is built from an input from Chrome, we cannot
enforce that a valid APN list will be provided. When a CustomAPNList is
used, but there is no DEFAULT APNs in the list, BuildDefaultApnTryList
will return an empty list. Return an error when this happens.

BUG=None
TEST=Enter an IA APN in the UI without any DEFAULT APNs.

Change-Id: I7128806bcfb5df8b9a4b32e9b3eb067f1e3fcc6e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4808878
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>