vm_tools: prevent s2idle on non-dgpu boards
Borealis has --s2idle enabled always, despite that 1) it is only
necessary for dgpu devices, and 2) it appears to break vsock currently.

We workaround this by refusing to add the arguments if there isn't a
dgpu present.

Fixing the issue with vsock+s2idle will be left for the future.

BUG=b:289456807
TEST=emerge, deploy, suspend_stress_test

Change-Id: I60832717a40d09e6b016822decdb5c90505c112f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4809835
Commit-Queue: Nic Hollingum <hollingum@google.com>
Reviewed-by: Grzegorz Jaszczyk <jaszczyk@google.com>
Tested-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Lucy Qu <lqu@google.com>