vm_tools: concierge: send vm_type to cicerone.
cicerone tracks vm_type but guesses it with a set of heuristics which
are different from the ones used by concierge (and ash).

Change it so that when concierge notifies cicerone of a VM starting, it
includes its (concierge's) value of vm_type.

Prefer vm_tools.apps.VmType per
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4369238

- system_api: add a system_api-vm_applications-goprotos target.
- NotifyVmStartedRequest message: add a vm_type field.
- ClassifyVm: add heuristic for bruschetta based on VM name.

BUG=b:294316866
TEST=cros-workon-atlas start vm_host_tools system_api chunnel crostini_client; emerge-atlas, cros deploy, manually install bruschetta

Change-Id: I2aaef4794b00d6d8d6887cc90efe65c8f9bc913c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4805429
Commit-Queue: Emil Mikulic <easy@google.com>
Reviewed-by: Fergus Dall <sidereal@google.com>
Tested-by: Emil Mikulic <easy@google.com>