fbpreprocessor: add session state manager
Monitor the StateChanged D-Bus signal from SessionManager to know when
the user logs in or out. When that happens, use RetrievePrimarySession
to know which user logged in (to check for permissions in the future)
and the path to the daemon-store that will be used to write firmware
dumps to disk (encrypted).

See design docs at:
go/cros-wifi-fw-dumps-architecture
go/cros-fbpreprocessord-dd

BUG=b:291346820
TEST=cros-workon-${BOARD} start fbpreprocessord
TEST=emerge-${BOARD} fbpreprocessord
TEST=cros deploy ${DUT} fbpreprocessord and push SessionManager.conf to
DUT
TEST=verify log messages when the user logs in/out.

Change-Id: I63e74712d910d0d5ef85104e4a7f1936db878b98
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4836708
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Reviewed-by: David Ruth <druth@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>