missive: Degradation improvement.
When degradation encounters a lower-priority queue with non-empty
writeable file, it was not erasing it, thus preventing higher-priority
queue from writing data.
With this change such a file will be closed and switched to a new one,
and the former file will be erased.

Added tests that verify that and changed the failed degradation tests to
reflect the new state.

Also, replaced std::ceil with UmaCeil function: it appears the former
produced 0 in case like std::ceil(192uL / 1024uL), because it first
does integer division and only then converts to floating and applies
std::ceil. We expected it to round up to 1.
UmaCeil(192uL, 1024uL) returns 1, as required.

BUG=b:281851540
TEST=cros_sdk cros_workon_make --board=${BOARD} missive --test
TEST=Manual test on VM

Change-Id: If63633a6257913c5170fbb942783f53550bbc7c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4811096
Commit-Queue: Josh Hilke <jrhilke@google.com>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Josh Hilke <jrhilke@google.com>