cryptohome: Support relabel for ephemeral sessions
This enhances the relabel operation to work with ephemeral sessions, and
also makes it work better with non-ephemeral ones by not discarding any
verifiers.

While there isn't a USS to act as a "transaction point" for ephemeral
users, the objects are also all in memory so it's pretty easy to do. It
just requires enhancing the verifier support to allow verifiers to be
extracted and renamed, rather than only added or removed.

BUG=b:292543364
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I4f5b2fbe80290347200230908e573d88a417b8dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4811052
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: George Engelbrecht <engeg@google.com>