cryptohome: Implement ModifyAuthFactorIntents
This CL implements ModifyAuthFactorIntents request. The intents in
the request are the proposed intents that should be enabled. Any
intent that needs to be enabled will be checked to be supported and
configurable before being enabled by this request. This change also
removes the config file flag that could have been used to enable
kDecrypt intent for Fingerprint. This intent can now be enabled via
ModifyAuthFactorIntents.

BUG=b:281878872
TEST=FEATURES=test emerge-amd64-generic cryptohome

Change-Id: I1f0260c9adc1de25384a41c95183e55bab108ea2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4812076
Reviewed-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: Behnood Momenzadeh <behnoodm@chromium.org>
Tested-by: Behnood Momenzadeh <behnoodm@chromium.org>