libec: biod: Deprecate FpPreloadTemplateCommand
The preload flow is going to be replaced by the template unlock flow.
It is not backward compatible but the whole feature hasn't launched so
it's fine. We'll first remove impl of FP preload template command as
FPMCU will remove it soon, then support the unlock template command and
migrate caller to the unlock flow, lastly we'll remove all references of
FP preload template commands.

BUG=b:290989633
TEST=libec biod unittests

Change-Id: I014f440498ded3b513616c686bfbee539a43b224
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4829405
Reviewed-by: Patryk Duda <patrykd@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>