patchpanel: add the DBUs interface for Bruschetta VM
This CL adds the DBus interface for starting and stopping a Bruschetta
VM. The pipeline includes patchpanel::Client, DBus proxy,
patchpanel::PatchpanelAdaptor, and patchpanel::Manager.

Note that this CL doesn't contain the implementation of the core
logic, and only put the simple argument and response for the Startup
method.

BUG=b:279994478
TEST=FEATURES=test chromeos-base/system_api patchpanel \
     patchpanel-client dns-proxy

Change-Id: Ief2f7395c703f0da9e8bfad8824948d03cc57f0c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4812856
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>