patchpanel: redact the client hostname from dnsmasq's log
If a DHCP client connected to a Chromebook's hotspot advertises a
hostname, that hostname is recorded by dnsmasq in /var/log/net.log
given the current syslog configuration. The host should be treated as
PII, so we should not record it at net.log.

Before parsing the hostname by "DHCPACK" log, the hostname only
appears with "client provides name: <hostname>". So This CL redacts
the client hostname by the steps:
1. Skip the log with "client provides name"
2. Get the hostname from the log: "DHCPACK"
3. Replace all the known hostname to "<redacted>"

This CL also fixes the bug that doesn't ignore the tokens prior
"DHCPACK" when parsing the MAC address and client hostname.

BUG=b:293403757
TEST=FEATURES=test emerge-brya patchpanel
TEST=Connect the hotspot and see the client hostname is redacted

Change-Id: I675ac72ec6c7cb1a84b23787e10e600b0117fc10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4812325
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>