camera: Fix effects pipeline hang caused by model change
Changing the segmentation model value triggers effects pipeline to reset
internally, and it corrupts the pipeline due to a closing failure. To
work around this, since we only use Auto and HD models where Auto is
actually resolved to HD internally, we resolve the Auto value earlier in
the stream manipulator to avoid model value changes to the pipeline.

BUG=b:296919762, b:296392886
TEST=Powerwash, start Meet and toggle VC effects.

Change-Id: I60d7e07e2b7bc5d69ee40db527f2a9c8f4c4eb40
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4812562
Tested-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Ren-Pei Zeng <kamesan@chromium.org>
Reviewed-by: Ricky Liang <jcliang@chromium.org>