cryptohome: Create EncryptedUss
Changes to extract the basic code for reading the raw encrypted USS
structure without decrypted it into a standalone class. This mostly just
extracts a bunch of that code from the existing UserSecretStash class
and switches existing auth session and UDA code that's accessing the raw
structure to instead use EncryptedUss.

For more detail about the eventual intended final state of the USS
classes see the linked bug and the doc that links to.

BUG=b:295898583
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ic2f1634686dfcf698b116d174625fd6d72cb4587
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4814310
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>