fbpreprocessor: add input manager
Make fbpreprocessord listen to the DebugDumpCreated D-Bus signal that is
emitted by crash-reporter when it has written a new firmware dump to
disk. When fbpreprocessord is notified, it starts the pseudonymization
of the firmware dump.
Once the pseudonymization is completed, the input firmware dump is
deleted. The firmware dumps are also deleted when the user logs in.

See design docs at:
go/cros-wifi-fw-dumps-architecture
go/cros-fbpreprocessord-dd

BUG=b:291346795
TEST=cros-workon-${BOARD} start fbpreprocessord
TEST=FEATURES=test emerge-${BOARD} fbpreprocessord
TEST=on DUT test described at http://gpaste/6142262005202944

Cq-Depend: chromium:4919988
Change-Id: I00775a652d2ab69b12d893283f25fa1d75693008
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4836711
Reviewed-by: David Ruth <druth@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>