fbpreprocessor: make expiration period configurable
By default, processed firmware dumps will be deleted after 30 minutes.
To make it possible to test the feature (by creating files and observing
that they do get automatically deleted from disk) we add the ability to
reduce the default expiration period. Tests will be able to set the
expiration period to, say, 30s, and will only have to wait for that
small amount of time.
Since the parameter is set from the arguments passed to the job, make
sure that the duration can't be extended beyond the default 30 minutes.

BUG=b:291346953
TEST=emerge-${BOARD} fbpreprocessord
TEST=cros deploy ${DUT} fbpreprocessord
TEST=start fbpreprocessord FILE_EXPIRATION_SECS=30 (on DUT)

Change-Id: I823444493e92b9bc722457c3da0d5f0886ab3c35
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4837554
Reviewed-by: David Ruth <druth@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>