vm_tools: concierge Implements the Balloon Class
A Balloon represents a single VM's balloon and handles all resizes of
that balloon. The balloon instance will automatically check for and
report balloon stalls through a callback. The balloon can be configured
to run blocking calls to crosvm on a separate sequence if necessary.

BUG=b:254164308
TEST=cros_run_unit_tests --board=BOARD --packages=vm_host_tools

Change-Id: I7b6480fc9ed0f19d325728f13add0a1fc09539b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4243468
Tested-by: Kameron Lutes <kalutes@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Reviewed-by: Keiichi Watanabe <keiichiw@chromium.org>
Commit-Queue: Kameron Lutes <kalutes@chromium.org>
Reviewed-by: David Stevens <stevensd@chromium.org>