lorgnette: Make JPEG output streamable
Lorgnette currently uses libjpeg's optimize_coding setting.  This causes
libjpeg to buffer the entire image in memory and write the final JPEG
when finalize is called.  For the current push-based GetNextImage, this
is not a problem because lorgnette directly writes the output to a
client-provided fd.  The Scan app waits for the scan to be fully done
before reading from the fd, so it never notices the difference between
progress messages and image content.

The new ReadScanData method will have the caller pulling data instead,
with image chunks returned directly via d-bus.  If the image data is
written all at the end, this produces a bunch of empty messages followed
by a final giant d-bus response.  In the worst case, the final message
might even exceed the maximum allowed size of a d-bus message.  To avoid
this behavior, encoded JPEG data needs to be generated in a single pass,
which means optimize_coding has to be disabled and bytes have to be
flushed to the output buffer after encoding each chunk of scanner data.

Unfortunately, disabling optimize_coding increases the file size by
10-25% across a set of test pages and devices.  To make up for the size
increase, also lower the quality setting.  The new setting is 90 for
lower resolution images and 85 for 300dpi and up.

Both changes together produce file sizes 0-15% smaller than the
originals for the same set of test documents and scanners.  Manual
inspection of the outputs didn't reveal any increase in visible JPEG
artifacts with text, flat graphics, or photographs at 100 or 300 dpi.

BUG=b:297443322
TEST=Compare before and after outputs from the Scan app

Cq-Depend: chromium:4885986
Change-Id: Ic69b374376442f4804b46a6dffc5861daf79b218
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4885023
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>