power: Clear Charge Limit display state on unplug
We are incorrectly clearing the display state of Charge Limit only when
the charge drops below the hold percent while plugged in. This changes
that to also clear the display state of Charge Limit on unplug. This
change also adds a unittest for this case.

BUG=b:296006442
TEST=manual
Wait for the device to be above 80% charge
In the shell, run the following:
echo 1 > /var/lib/power_manager/charge_limit_enabled
restart powerd
Unplug the system and check that the display battery % can change
TEST=unittests

Change-Id: I5e49aebfc9cd5f644c36e8c2bbe6d453a6f20bd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4820573
Auto-Submit: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Ivan Chen <yulunchen@google.com>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>