crash: Join session keyring earlier.
We need the keyring to be unlocked in order to write crash_reporter
collection errors to daemon-store.

However, do *not* unconditionally join it; if we already joined the
keyring (e.g. we are a descendent of init), joining it a second time
will result in the kernel silently creating a new keyring with the same
name, but without the permissions to access the directories we need.

BUG=b:297992960
TEST=tast run puff_dut crash.User* crash.KernelWarning* \
     crostini.CrashReporter.*_stable \
     ui.ChromeCrashLoggedIn.gpu_process_crashpad

Change-Id: I5d78ec583cd79787ae231ed9047e16aa0bebfd07
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4824088
Tested-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>