lorgnette: Start and stop scan jobs
StartPreparedScan initiates a scan with the current option settings.
CancelScan takes a handle returned from StartPreparedScan and cancels
it.  This CL implements the pre-checks and requests the SANE device to
start/stop the scan.

Because ReadScanData is not yet implemented, there is nothing useful
that can be done with the scan yet, but this gets the basic job handle
framework in place.

BUG=b:297443322
TEST=CQ

Change-Id: I0c000a650bd7c537dfd3e0e30272289cc9cfce51
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4836734
Reviewed-by: Clayton Whitelaw <cawhitelaw@google.com>
Reviewed-by: Nathan Muggli <nmuggli@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>