power: Only overwrite display battery percent while within range
This changes Adaptive Charging and Charge Limit to only overwrite the
display battery percentage when within the charge range specified for
the battery sustainer. This is added as an extra check to ensure we
don't accidentally show the wrong battery charge to the user.

BUG=b:296006442
TEST=unittests

Change-Id: Id01e05b09703d24ed1aec66230522a789afb4a2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4826163
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>