lorgnette: Report supported image formats
One of the fields in ScannerInfo gives a list of image formats that can
be retrieved from a given scanner.  lorgnette will eventually support
passing through advanced formats directly from the scanner, but for now
just declare support for the JPG and PNG files it can already generate
internally.

In the process of adding constants for the new MIME types, also refactor
constants.h to properly declare an extern variable instead of relying on
the previous textual inclusion of a static variable.

BUG=b:297443322
TEST=lorgnette_cli discover --show_details

Change-Id: I55e937e219eb16e44d7b582833929c8c7371daef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4832136
Reviewed-by: Mason Wilde <masonwilde@google.com>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Benjamin Gordon <bmgordon@chromium.org>