patchpanel: arc: delay "arc0" Device creation
To swap Device with ArcService::ArcDevice in ArcService, it is required
to know the name of the virtual interface created on the host for the
ARC device. For ARCVM this requires actually creating the tap device in
the host kernel, otherwise the name cannot be predicted in advance
depending on which other guest VM is running.

This information is not obtained until ARCVM actually starts and
ArcService::Start() is run. Therefore it is necessary to delay the
creation of the "arc0" Device to ::Start(), and to also destroy it in
::Stop().

BUG=b:293954997
TEST=FEATURES=test emerge-$BOARD patchpanel

Change-Id: Ie04c96e83c8578d4c50c0ac1a11dbb6be2c6f396
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4834901
Reviewed-by: Ningyuan Wang <ningyuan@google.com>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>