sepolicy: Add policy for fpstudy
This is adds support for the fingerprint study testing/diagnostics tool.
This tool is not shipped in production images, but does need to
be installed to rootfs for custom unsigned images, for fingerprint
testing.

The tmpfiles config was added in https://crrev.com/c/4060398,
but it did not setup the correct sepolicy. This resulted in
the stateful being wiped upon reboot, since systemd-tmpfiles
would yield errors on boot.

BUG=b:298314270
TEST=Flash and run image from https://crrev.com/c/4829626
     Ensure that rebooting doesn't cause a stateful reset.
TEST=/usr/bin/systemd-tmpfiles --create --remove --boot --prefix /home --prefix /var
     # Make sure no errors are related to fpstudy
TEST=Make sure /var/log/fpstudy-audit.log exists and is working.
TEST=Make sure /var/log/fingerprint dir has valid logs
TEST=Make sure /var/lib/fingers contains fingerprint captures
    after collecting a few.

Change-Id: Ibd13e153b5e41a26fdeda386a22894ab13eba50e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4829157
Tested-by: Craig Hesling <hesling@chromium.org>
Reviewed-by: Firas Sammoura <fsammoura@google.com>
Commit-Queue: Craig Hesling <hesling@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>