secagentd: Fix redaction delay
Currently there is a 2 second delay after each sign in for the
device user to be filled because of time it takes to determine if
the user is affiliated. (Only relevant on first sign in). This is
bad for redaction because if the username appears in the cmdline
it cannot be redacted because it is not yet known. Move the 2 second
delay after the username is added to redaction list.
Also change redaction message to differentiate between server redaction.

BUG=b:291642096
TEST=Unit tests pass

Change-Id: I6c60012f33ed147edd597e5c73d80ee93363eedb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4830593
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Commit-Queue: Ryan Borzello <rborzello@google.com>
Tested-by: Ryan Borzello <rborzello@google.com>