patchpanel: arc: regroup veth pair creation and destruction
This patch regroups together the creation and destruction of veth pairs
for the ARC container in the ARC Device Datapath configuration.

This regrouping facilitates migrating to ArcDevice by separating better
the inference of the virtual interface name on the host from the
Datapath configuration code.

One behavior difference is that ARC initial setup does not fail anymore
if the "arc0" veth device cannot be created for the container. Since
"arc0" is only used for VPN forwarding, ARC connectivity can still be
functional without it for all use cases without a VPN connection.

BUG=b:293954997
TEST=FEATURES=test emerge-$BOARD patchpanel

Change-Id: I134f44315e508d446bdb7d58628835dc97ba8402
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4834903
Tested-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Reviewed-by: Ningyuan Wang <ningyuan@google.com>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>